+++ src/trx/game/lua/assault.c
@@ -1,6 +1,7 @@
 #include <trx/config.h>
 #include <trx/config/types.h>
 #include <trx/game/const.h>
+#include <trx/game/game.h>
 #include <trx/game/gym.h>
 #include <trx/game/lua/common.h>
 
@@ -55,12 +56,76 @@ static bool M_RemoveAssaultTimeAtIndex(const int32_t idx)
     return true;
 }
 
-// trxc.assault_stats.record(time) -> bool
-static int M_L_AssaultStatsRecord(lua_State *const L)
+static GYM_TRACK_TYPE M_GetTrack(lua_State *const L)
+{
+    const int32_t raw_track = (int32_t)luaL_optinteger(L, 1, GYM_TRACK_ASSAULT);
+
+    switch (raw_track) {
+    case GYM_TRACK_ASSAULT:
+    case GYM_TRACK_QUAD:
+        return (GYM_TRACK_TYPE)raw_track;
+    default:
+        luaL_error(L, "Invalid track (expected Track.COURSE or Track.QUAD)");
+        return GYM_TRACK_ASSAULT;
+    }
+}
+
+static const char *M_GetTrackName(const GYM_TRACK_TYPE track)
+{
+    switch (track) {
+    case GYM_TRACK_ASSAULT:
+        return "Assault course";
+    case GYM_TRACK_QUAD:
+        return "Quad Bike";
+    default:
+        return "Unknown";
+    }
+}
+
+static void M_CheckTimerAvailable(
+    lua_State *const L, const GYM_TRACK_TYPE track)
+{
+    if (!Game_IsInGym()) {
+        luaL_error(
+            L, "%s timer is only available in gym levels",
+            M_GetTrackName(track));
+    }
+
+    if (!Gym_TrackManager_HasStats(track)) {
+        luaL_error(L, "%s timer unavailable", M_GetTrackName(track));
+    }
+}
+
+static int M_L_AssaultStart(lua_State *const L)
+{
+    const GYM_TRACK_TYPE track = M_GetTrack(L);
+    M_CheckTimerAvailable(L, track);
+    Gym_TrackManager_Start(track);
+    return 0;
+}
+
+static int M_L_AssaultStop(lua_State *const L)
+{
+    const GYM_TRACK_TYPE track = M_GetTrack(L);
+    M_CheckTimerAvailable(L, track);
+    Gym_TrackManager_Stop(track);
+    return 0;
+}
+
+static int M_L_AssaultReset(lua_State *const L)
+{
+    const GYM_TRACK_TYPE track = M_GetTrack(L);
+    M_CheckTimerAvailable(L, track);
+    Gym_TrackManager_Reset(track);
+    return 0;
+}
+
+static int M_L_AssaultRecord(lua_State *const L)
 {
     if (!Gym_TrackManager_HasStats(GYM_TRACK_ASSAULT)) {
         return luaL_error(L, "Assault stats unavailable");
     }
+
     const float time = (float)luaL_checknumber(L, 1);
     if (time <= 0.0f) {
         return luaL_error(L, "Time must be > 0");
@@ -71,12 +136,12 @@ static int M_L_AssaultStatsRecord(lua_State *const L)
     return 1;
 }
 
-// trxc.assault_stats.remove(index) -> bool
-static int M_L_AssaultStatsRemove(lua_State *const L)
+static int M_L_AssaultRemoveRecord(lua_State *const L)
 {
     if (!Gym_TrackManager_HasStats(GYM_TRACK_ASSAULT)) {
         return luaL_error(L, "Assault stats unavailable");
     }
+
     const int64_t index_1 = luaL_checkinteger(L, 1);
     if (index_1 < 1 || index_1 > MAX_ASSAULT_TIMES) {
         return luaL_error(
@@ -89,8 +154,7 @@ static int M_L_AssaultStatsRemove(lua_State *const L)
     return 1;
 }
 
-// trxc.assault_stats.list() -> { { time = float, attempt_num = int }, ... }
-static int M_L_AssaultStatsList(lua_State *const L)
+static int M_L_AssaultListRecords(lua_State *const L)
 {
     if (!Gym_TrackManager_HasStats(GYM_TRACK_ASSAULT)) {
         return luaL_error(L, "Assault stats unavailable");
@@ -118,18 +182,34 @@ static int M_L_AssaultStatsList(lua_State *const L)
     return 1;
 }
 
-void LUA_CreateAssaultStats(lua_State *const L)
+void LUA_CreateAssault(lua_State *const L)
 {
     lua_getglobal(L, "trxc");
     lua_newtable(L);
 
-    lua_pushcfunction(L, M_L_AssaultStatsRecord);
-    lua_setfield(L, -2, "record");
-    lua_pushcfunction(L, M_L_AssaultStatsRemove);
-    lua_setfield(L, -2, "remove");
-    lua_pushcfunction(L, M_L_AssaultStatsList);
-    lua_setfield(L, -2, "list");
+    lua_pushcfunction(L, M_L_AssaultStart);
+    lua_setfield(L, -2, "start");
+    lua_pushcfunction(L, M_L_AssaultStop);
+    lua_setfield(L, -2, "stop");
+    lua_pushcfunction(L, M_L_AssaultReset);
+    lua_setfield(L, -2, "reset");
 
-    lua_setfield(L, -2, "assault_stats");
+    lua_newtable(L);
+    lua_pushcfunction(L, M_L_AssaultRecord);
+    lua_setfield(L, -2, "record");
+    lua_pushcfunction(L, M_L_AssaultRemoveRecord);
+    lua_setfield(L, -2, "remove");
+    lua_pushcfunction(L, M_L_AssaultListRecords);
+    lua_setfield(L, -2, "list");
+    lua_setfield(L, -2, "stats");
+
+    lua_newtable(L);
+    lua_pushinteger(L, GYM_TRACK_ASSAULT);
+    lua_setfield(L, -2, "COURSE");
+    lua_pushinteger(L, GYM_TRACK_QUAD);
+    lua_setfield(L, -2, "QUAD");
+    lua_setfield(L, -2, "Track");
+
+    lua_setfield(L, -2, "assault");
     lua_pop(L, 1);
 }